Rename: question api 구조 수정
