--- src/api/services/question/random/useRandomQuestion.ts
@@ -1,10 +1,0 @@
-import { useQuery } from '@tanstack/react-query'
-
-import { fetchRandomQuestion } from '.'
-
-export const useRandomQuestion = () => {
-  return useQuery({
-    queryKey: ['randomQuestion'],
-    queryFn: fetchRandomQuestion,
-  })
-}
